+++ RELEASE_NOTES.md
@@ -145,7 +145,7 @@ compiler/runtime.
 * ADDED: Bash shell script to create a JIProlog `.jip` file with the Logtalk
 compiler/runtime.
 
-* ADDED: Bash shell script to create a ECLiPSe `.asm` file with the Logtalk
+* ADDED: Bash shell script to create a ECLiPSe `.eco` file with the Logtalk
 compiler/runtime.
 
 * UPDATED: The Windows installation script to also detect ECLiPSe 7.1